requirements.txt
bleach==2.0.0
diff-match-patch==20121119
Django==1.9.8
django-appconf==1.0.2
django-appcore==0.0.9
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==0.6.0
django-ipware==1.1.6
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
et-xmlfile==1.0.1
future==0.16.0
html5lib==0.999999999
httplib2==0.9.2
jdcal==1.3
Markdown==2.6.11
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
Pygments==2.2.0
PyMySQL==0.7.11
PyYAML==3.12
six==1.11.0
tablib==0.12.1
timeago==1.0.8
unicodecsv==0.14.1
webencodings==0.5.1
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
